Allow for multi use endpoints by accepting space suffixes on requests
A more complete solution to this would require revamping the hash structure that stores the individual HTTP verb blocks within each request path - a not insignificant task!

In place of that level of complexity, this change allows users to add one or more spaces onto the end of the paths within their spec to make them "unique". This has the advantage of being simple to implement with very minimal code change.